Upgrade require_all to 2.0
This is in response to a breaking change introduced by require_all 2.0

When using require_all 2.0, the gem runs into an error when requiring
`lib/formulas`, as each file is loaded in alphabetical order, which
means that some formulas that are subclasses of Formula are loaded
before `lib/formulas/formula.rb`, resulting in a RequireAll::LoadError
for the uninitialized constant Formula.

This is easily fixed by requiring `lib/formulas/formula.rb` before
`lib/formulas`. Also, explicit versioning of require_all is introduced
to prevent future errors of this kind.